IntimateTales_Roleplay/src/interactive_tools/decision_maker.py
```python
from ..utils import utils
from ..utils import constants

class DecisionMaker:
    def __init__(self):
        self.decision_history = []

    def make_decision(self, decision_options):
        """
        Function to make a decision based on the provided options.
        """
        # Validate the decision options
        if not utils.validate_decision_options(decision_options):
            raise ValueError("Invalid decision options provided.")

        # Make a decision
        decision = utils.make_decision(decision_options)

        # Update the decision history
        self.update_decision_history(decision)

        return decision

    def update_decision_history(self, decision):
        """
        Function to update the decision history.
        """
        # Validate the decision
        if not utils.validate_decision(decision):
            raise ValueError("Invalid decision provided.")

        # Update the decision history
        self.decision_history.append(decision)

    def get_decision_history(self):
        """
        Function to get the decision history.
        """
        return self.decision_history

    def clear_decision_history(self):
        """
        Function to clear the decision history.
        """
        self.decision_history = []
```